Split map component
The map component will be reused in the obsidian and web version.
Moved the map implementation to common
