+++ src/obsidian/DistortableImageOverlay.ts
