+++ src/obsidian/embedded_map.ts
