+++ src/obsidian/event.ts
